fix for issue #319

+++ eclipse_ws/biobank2/src/edu/ualberta/med/biobank/model/SiteClinicInfo.java
@@ -9,7 +9,7 @@ public class SiteClinicInfo implements ITableInfo {
 
     public int studies;
 
-    public int patients;
+    public Long patients;
 
     public int patientVisits;
